platzi-product-api-crud: pagination, sorting, searching, filter

+++ javascript/platzi-product-api-crud/main.js
@@ -1,4 +1,7 @@
-let currentPageOffset = 0;
+let currentPage = 1;
+const itemsPerPage = 12;
+let products = [];
+let categories = [];
 
 const loadingContainer = document.getElementById("loading");
 const noRecords = document.getElementById("no-records");
@@ -6,19 +9,50 @@ const modal = document.getElementById("myModal");
 const span = document.getElementsByClassName("close")[0];
 const modalBody = document.getElementById("modal-body");
 const paginationContainer = document.getElementById("pagination-container");
+const filterCategoryDropdown = document.getElementById("filterCategory");
+const sortDropdown = document.getElementById("sort");
+const sortPriceDropdown = document.getElementById("sortPrice");
 
 span.onclick = function () {
   modal.style.display = "none";
   modalBody.innerHTML = "";
 };
 
-fetchProducts(0);
-function fetchProducts(offset) {
+document.getElementById("search").addEventListener("input", function (e) {
+  handleSearch();
+});
+
+filterCategoryDropdown.addEventListener("change", function (e) {
+  handleFilterByCategory();
+});
+
+sortDropdown.addEventListener("change", function (e) {
+  handleSort();
+});
+
+sortPriceDropdown.addEventListener("change", function (e) {
+  handleSortPrice();
+});
+
+document.getElementById("reset").addEventListener("click", handleReset);
+function handleReset() {
+  document.getElementById("search").value = "";
+  filterCategoryDropdown.value = "";
+  sortDropdown.value = "";
+  sortPriceDropdown.value = "";
+  fetchProducts();
+}
+
+fetchProducts();
+
+//
+function fetchProducts() {
   loadingContainer.style.display = "block";
   const container = document.getElementById("product-container");
   container.innerHTML = "";
 
-  fetch(`https://api.escuelajs.co/api/v1/products?offset=${offset}&limit=10`)
+  let fetchUrl = `https://api.escuelajs.co/api/v1/products`;
+  fetch(fetchUrl)
     .then((response) => {
       if (!response.ok) {
         throw new Error("Failed to fetch products.");
@@ -27,11 +61,16 @@ function fetchProducts(offset) {
     })
     .then((res) => {
       loadingContainer.style.display = "none";
-      console.log(res);
-      let products = res;
-      displayProductsForPage(products);
-      renderPagination(100, offset);
-      currentPageOffset = offset;
+
+      products = res;
+      if (products.length === 0) {
+        noRecords.style.display = "block";
+      } else {
+        noRecords.style.display = "none";
+        renderPagination(products);
+        displayProductsForPage(currentPage, products);
+        updateCategoriesDropdown();
+      }
     })
     .catch((err) => {
       loadingContainer.style.display = "none";
@@ -39,36 +78,39 @@ function fetchProducts(offset) {
     });
 }
 
-function fetchCurrentPageProducts() {
-  fetchProducts(currentPageOffset);
-}
-
-function displayProductsForPage(products) {
+//
+function displayProductsForPage(page, productList) {
   const container = document.getElementById("product-container");
   container.innerHTML = "";
 
-  products.forEach((res, index) => {
+  const startIndex = (page - 1) * itemsPerPage;
+  const endIndex = startIndex + itemsPerPage;
+  const pageProducts = productList.slice(startIndex, endIndex);
+
+  pageProducts.forEach((res, index) => {
     const createDiv = document.createElement("div");
     createDiv.classList.add("card");
     createDiv.addEventListener("click", () => {
       handleModal(res);
     });
+    const imageSrc =
+      "https://imgs.search.brave.com/Rl9vjXH4729H-gE4Cz-KidigqEe0iE05CnRhUt2a2l0/rs:fit:860:0:0/g:ce/aHR0cHM6Ly93LmZv/cmZ1bi5jb20vZmV0/Y2gvOWIvOWJlNTQ3/MDY4MTNlZWEyNTg3/YzA1ZmU4YWJlMGJl/NzAuanBlZw";
+
     createDiv.innerHTML = `
-        <div class="title">${res.title}</div>
-        <div class="image-container">
-          <img src="https://imgs.search.brave.com/jBmM81B7TVdzHAJtxUZqFKPbB649UftZA7939rc4sK8/rs:fit:860:0:0/g:ce/aHR0cHM6Ly9jZG4u/cGl4YWJheS5jb20v/cGhvdG8vMjAxNy8x/MC8xMi8yMi8xNy9i/dXNpbmVzcy0yODQ2/MjIxXzY0MC5qcGc" alt="image" width="50" />
-        </div>
-        `;
+      <div class="image-container">
+        <img src="${imageSrc}" alt="image" width="50" />
+      </div>
+      <div class="title">${res.title}</div>`;
     container.appendChild(createDiv);
   });
 }
 
+//
 function handleModal(res) {
   modal.style.display = "block";
-  const modalBody = document.getElementById("modal-body");
   modalBody.innerHTML = `
         <div class='modal-image-container'>
-          <img src="https://imgs.search.brave.com/jBmM81B7TVdzHAJtxUZqFKPbB649UftZA7939rc4sK8/rs:fit:860:0:0/g:ce/aHR0cHM6Ly9jZG4u/cGl4YWJheS5jb20v/cGhvdG8vMjAxNy8x/MC8xMi8yMi8xNy9i/dXNpbmVzcy0yODQ2/MjIxXzY0MC5qcGc" alt="image" width="50" />
+            <img src="https://imgs.search.brave.com/Rl9vjXH4729H-gE4Cz-KidigqEe0iE05CnRhUt2a2l0/rs:fit:860:0:0/g:ce/aHR0cHM6Ly93LmZv/cmZ1bi5jb20vZmV0/Y2gvOWIvOWJlNTQ3/MDY4MTNlZWEyNTg3/YzA1ZmU4YWJlMGJl/NzAuanBlZw"; alt="image" width="50" />
         </div>
         <div class='modal-content-container'>
             <div>${res.title}</div>
@@ -77,64 +119,70 @@ function handleModal(res) {
             <div>Category: ${res.category.name}</div>
             <button onclick=deleteProduct(${res.id})>Delete</button>
             <button onclick=updateProduct(${res.id})>Update</button>    
-        </div>
-    `;
+        </div>`;
 }
 
+//
 function updateProduct(id) {
   window.location.href = `./update/index.html?id=${id}`;
 }
 
+//
 function deleteProduct(id) {
-  fetch(`https://api.escuelajs.co/api/v1/products/${id}`, {
-    method: "DELETE",
-  })
-    .then((res) => res.json())
-    .then((response) => {
-      modal.style.display = "none";
-      fetchCurrentPageProducts();
+  const response = confirm("Are you sure you want to delete this product?");
+  if (response) {
+    fetch(`https://api.escuelajs.co/api/v1/products/${id}`, {
+      method: "DELETE",
     })
-    .catch((error) => {
-      console.log(error);
-    });
+      .then((res) => res.json())
+      .then((response) => {
+        modal.style.display = "none";
+        fetchProducts();
+        document.getElementById("search").value = "";
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  }
 }
 
-function renderPagination(total, offset) {
-  const itemsPerPage = 10;
-  const pageCount = Math.ceil(total / itemsPerPage);
-  const paginationContainer = document.getElementById("pagination-container");
+//
+function renderPagination(productList) {
+  const pageCount = Math.ceil(productList.length / itemsPerPage);
   paginationContainer.innerHTML = "";
 
   if (pageCount > 1) {
     const prevButton = createPaginationButton("Previous");
     prevButton.addEventListener("click", () => {
-      const newOffset = Math.max(0, offset - itemsPerPage);
-      fetchProducts(newOffset);
+      if (currentPage > 1) {
+        currentPage--;
+        displayProductsForPage(currentPage, productList);
+        updatePaginationButtons();
+      }
     });
-    prevButton.disabled = offset === 0;
+    prevButton.disabled = currentPage === 1;
 
     const nextButton = createPaginationButton("Next");
     nextButton.addEventListener("click", () => {
-      const newOffset = Math.min(total - itemsPerPage, offset + itemsPerPage);
-      fetchProducts(newOffset);
+      if (currentPage < pageCount) {
+        currentPage++;
+        displayProductsForPage(currentPage, productList);
+        updatePaginationButtons();
+      }
     });
-    nextButton.disabled = offset >= total - itemsPerPage;
+    nextButton.disabled = currentPage === pageCount;
 
     paginationContainer.appendChild(prevButton);
 
-    const currentPage = Math.floor(offset / itemsPerPage) + 1;
-    const startPage = Math.max(1, currentPage - 1);
-    const endPage = Math.min(pageCount, startPage + 3);
-    for (let i = startPage; i <= endPage; i++) {
+    for (let i = 1; i <= pageCount; i++) {
       const pageButton = createPaginationButton(i.toString());
+      pageButton.addEventListener("click", () => {
+        currentPage = i;
+        displayProductsForPage(currentPage, productList);
+        updatePaginationButtons();
+      });
       if (i === currentPage) {
         pageButton.classList.add("active");
-        pageButton.disabled = true;
-      } else {
-        pageButton.addEventListener("click", () => {
-          const newOffset = (i - 1) * itemsPerPage;
-          fetchProducts(newOffset);
-        });
       }
       paginationContainer.appendChild(pageButton);
     }
@@ -143,9 +191,124 @@ function renderPagination(total, offset) {
   }
 }
 
+//
+function updatePaginationButtons() {
+  const pageCount = Math.ceil(products.length / itemsPerPage);
+  const paginationButtons = document.querySelectorAll(".pagination-button");
+  paginationButtons.forEach((button, index) => {
+    if (index === 0) {
+      button.disabled = currentPage === 1;
+    } else if (index === paginationButtons.length - 1) {
+      button.disabled = currentPage === pageCount;
+    }
+    button.classList.remove("active");
+  });
+  paginationButtons[currentPage].classList.add("active");
+}
+
+//
 function createPaginationButton(text) {
   const button = document.createElement("button");
   button.textContent = text;
   button.classList.add("pagination-button");
   return button;
+}
+
+//
+function handleFilterByCategory() {
+  const selectedCategory = filterCategoryDropdown.value;
+  let filteredProducts = products;
+  if (selectedCategory !== "") {
+    filteredProducts = products.filter(
+      (product) => product.category.name === selectedCategory
+    );
+  }
+  currentPage = 1;
+  renderPagination(filteredProducts);
+  displayProductsForPage(currentPage, filteredProducts);
+}
+
+//
+function handleSort() {
+  const sortOption = sortDropdown.value;
+  let sortedProducts = [...products];
+  const selectedCategory = filterCategoryDropdown.value;
+
+  if (selectedCategory !== "") {
+    sortedProducts = sortedProducts.filter(
+      (product) => product.category.name === selectedCategory
+    );
+  }
+
+  if (sortOption === "a-z") {
+    sortedProducts.sort((a, b) => a.title.localeCompare(b.title));
+  } else if (sortOption === "z-a") {
+    sortedProducts.sort((a, b) => b.title.localeCompare(a.title));
+  }
+
+  currentPage = 1;
+  renderPagination(sortedProducts);
+  displayProductsForPage(currentPage, sortedProducts);
+}
+
+//
+function handleSortPrice() {
+  const sortPriceOption = sortPriceDropdown.value;
+  let sortedProducts = [...products];
+  const selectedCategory = filterCategoryDropdown.value;
+
+  if (selectedCategory !== "") {
+    sortedProducts = sortedProducts.filter(
+      (product) => product.category.name === selectedCategory
+    );
+  }
+
+  if (sortPriceOption === "low-to-high") {
+    sortedProducts.sort((a, b) => a.price - b.price);
+  } else if (sortPriceOption === "high-to-low") {
+    sortedProducts.sort((a, b) => b.price - a.price);
+  }
+
+  currentPage = 1;
+  renderPagination(sortedProducts);
+  displayProductsForPage(currentPage, sortedProducts);
+}
+
+//
+function handleSearch() {
+  const searchInput = document.getElementById("search").value.toLowerCase();
+  const filteredProducts = products.filter((product) =>
+    product.title.toLowerCase().includes(searchInput)
+  );
+  currentPage = 1;
+  renderPagination(filteredProducts);
+  displayProductsForPage(currentPage, filteredProducts);
+
+  if (filteredProducts.length === 0) {
+    noRecords.style.display = "block";
+  } else {
+    noRecords.style.display = "none";
+  }
+}
+
+//
+function updateCategoriesDropdown() {
+  categories = products.map((product) => product.category.name);
+  const uniqueCategories = Array.from(new Set(categories));
+
+  filterCategoryDropdown.innerHTML =
+    "<option value='' selected>Select Category</option>";
+
+  uniqueCategories.forEach((category) => {
+    const option = document.createElement("option");
+    option.value = category;
+    option.textContent = category;
+    filterCategoryDropdown.appendChild(option);
+  });
+
+  if (filteredProducts.length === 0) {
+    noRecords.style.display = "block";
+  } else {
+    noRecords.style.display = "none";
+  }
 }
